requirements.txt
# ============================================
# 로일(LoIl) - Python Dependencies
# ============================================

# ============================================
# Discord Bot
# ============================================
discord.py==2.3.2
discord-ext-menus==1.0.0a34

# ============================================
# Database
# ============================================
# MySQL (추천 - 개발자 경험)
pymysql==1.1.0
mysqlclient==2.2.1

# 또는 PostgreSQL
# psycopg2-binary==2.9.9

# ORM & Migration
SQLAlchemy==2.0.23
alembic==1.13.1

# ============================================
# AI
# ============================================
# Gemini (Google) - 메인 AI
google-generativeai==0.3.2

# Groq - 백업 AI
groq==0.4.2

# OpenAI (선택사항)
# openai==1.6.1

# ============================================
# APIs
# ============================================
requests==2.31.0
aiohttp==3.9.1

# ============================================
# Google Sheets
# ============================================
gspread==5.12.3
oauth2client==4.1.3
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# ============================================
# 카카오톡 (차후 추가)
# ============================================
# kakao-sdk==1.0.0

# ============================================
# 환경변수
# ============================================
python-dotenv==1.0.0

# ============================================
# 데이터 검증
# ============================================
pydantic==2.5.3
pydantic-settings==2.1.0

# ============================================
# 유틸리티
# ============================================
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0

# ============================================
# 보안 (API 키 암호화)
# ============================================
cryptography==41.0.7
bcrypt==4.1.2

# ============================================
# 로깅
# ============================================
loguru==0.7.2

# ============================================
# 테스트
# ============================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# ============================================
# 코드 품질
# ============================================
black==23.12.1
ruff==0.1.9
flake8==6.1.0
mypy==1.7.1

# ============================================
# 개발 도구
# ============================================
ipython==8.19.0
rich==13.7.0

# ============================================
# 웹 (선택사항 - 향후)
# ============================================
# fastapi==0.108.0
# uvicorn==0.25.0
# jinja2==3.1.2

# ============================================
# 모니터링 (선택사항)
# ============================================
# sentry-sdk==1.39.2

# ============================================
# Redis (캐싱, 선택사항)
# ============================================
# redis==5.0.1
# hiredis==2.3.2

# ============================================
# 이미지 처리 (차트 생성용)
# ============================================
# pillow==10.1.0
# matplotlib==3.8.2

# ============================================
# 설치 방법
# ============================================
#
# 기본 설치:
#   pip install -r requirements.txt
#
# 개발 환경:
#   pip install -r requirements.txt
#
# 프로덕션 환경:
#   pip install -r requirements.txt --no-dev
#
# 가상환경 사용 (권장):
#   python -m venv venv
#   source venv/bin/activate  # Linux/Mac
#   venv\Scripts\activate     # Windows
#   pip install -r requirements.txt
#
# ============================================
